test(integration): clean up docker containers on SIGINT/SIGTERM
Jest's globalTeardown only runs on a clean exit, so Ctrl+C or a CI
cancel would leak the anvil/voltaire containers and the docker network
until the next successful run. Extract the teardown logic into a shared
helper and register SIGINT/SIGTERM handlers in globalSetup that invoke
it and re-raise the signal.
